Answer the pre-merge table's two warnings
Developer Documentation: the timer table still said the job ceiling was
60 minutes and the sizing paragraph still described 60 as ten minutes
above a 50-minute requirement. The requirement has three terms and the
ceiling equals it: a 2,700 s watchdog, 300 s of measured work outside
it, and a 900 s margin, which is 65 minutes.

Testing (Property / Proof): required_ceiling takes any sequence of
budgets, and both lanes here run the coverage action once with the same
watchdog, so the example the contract asserts against would pass a
reading that returned its first argument, multiplied by the step count,
or dropped either constant. Three Hypothesis properties pin the shape
instead: every term is added for any number of steps, adding a step
never lowers the requirement, and the result never falls below the sum
it must contain plus the margin.

Proved by mutation: dropping CEILING_MARGIN_SECONDS fails two of the
three, and taking the maximum rather than the sum fails two.

The module reached 652 lines against AGENTS.md's 400-line limit, so the
lane reading moves to _timeout_lane_support.py and the assertions stay
in the contract.
